requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
SQLAlchemy==2.0.12
typing_extensions==4.5.0
Werkzeug
zipp==3.15.0